n8n/agent-usecases/pyproject.toml
[project]
name = "agent-use-cases"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "selenium>=4.33.0",
    "streamlit>=1.45.1",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "beautifulsoup4>=4.13.4",
    "langchain-openai>=0.3.24",
    "gspread>=6.2.1",
    "nest-asyncio>=1.6.0",
    "langchain-text-splitters>=0.3.8",
    "langchain-community>=0.3.25",
    "unstructured>=0.17.2",
    "markdown>=3.8.2",
    "langchain-pinecone>=0.2.8",
    "pq>=1.9.1",
    "langchain-mcp-adapters>=0.1.7",
    "faker>=37.4.0",
    "pydub>=0.25.1",
    "yt-dlp>=2025.6.30",
    "psycopg[binary]>=3.2.9",
]
